Guarantees itemSets are always Sets
Ensures that `itemSets` are always instances of the `Set` object
before attempting to use them as such.

This prevents potential errors and ensures consistency in how grouped
items are handled. It adds a check to ensure that `itemSets` is
either undefined or an instance of `Set` before attempting to use
it. If it is not, it initializes it as a new `Set`.